Add basic authentication on request.
Adds a global preprocessor to every HTTP request that handles HTTP basic
authentication if, and only if, it is there. If the authentication
succeeds, it pushes the authenticated player object to Flask's request
globals. If it fails, it cancels request processing and returns a 401
status. If no Authorization header is present, it sets the player object
in Flask's globals to None.

Clients can pass authentication details whenever they feel like, and the
authorization methods of the data objects can return 401 when they find
need to and tell the client to add authentication.