Store username and password seperately.
We changed this to the pre-generated Authorization header because the
username/password thing in XHR only kicks in after receiving a 401
response, leading to one 401 response before every successful response.
Sending the Authorization header directly removed half of the HTTP
requests.

However, it turns out that, setting username and password in XHR injects
the credentials into the browser's session store, making it valid also
for all following browser requests to the site, like for images in case
they produce a 401 status.

This hopefully completes #42.